Refactor Whisper resampling helper
