+++ src-tauri/src/whisper/recorder.rs
@@ -1,5 +1,6 @@
 //! Native audio recording with streaming support using cpal
 
+use super::audio::resample;
 use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
 use cpal::Stream;
 use std::sync::atomic::{AtomicBool, AtomicU32, AtomicUsize, Ordering};
@@ -242,41 +243,6 @@ impl AudioRecorder {
         }
         samples.len() as f32 / input_rate as f32
     }
-}
-
-impl Default for AudioRecorder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Simple linear interpolation resampling
-fn resample(samples: &[f32], from_rate: u32, to_rate: u32) -> Vec<f32> {
-    if from_rate == to_rate || samples.is_empty() {
-        return samples.to_vec();
-    }
-
-    let ratio = from_rate as f64 / to_rate as f64;
-    let new_len = (samples.len() as f64 / ratio).ceil() as usize;
-    let mut resampled = Vec::with_capacity(new_len);
-
-    for i in 0..new_len {
-        let pos = i as f64 * ratio;
-        let idx = pos.floor() as usize;
-        let frac = pos - pos.floor();
-
-        let sample = if idx + 1 < samples.len() {
-            samples[idx] * (1.0 - frac as f32) + samples[idx + 1] * frac as f32
-        } else if idx < samples.len() {
-            samples[idx]
-        } else {
-            0.0
-        };
-
-        resampled.push(sample);
-    }
-
-    resampled
 }
 
 #[cfg(test)]
