+++ src-tauri/src/whisper/transcriber.rs
@@ -1,5 +1,6 @@
 //! Local whisper transcription using whisper-rs
 
+use super::audio::resample;
 use hound;
 use serde::{Deserialize, Serialize};
 use std::path::{Path, PathBuf};
@@ -182,35 +183,6 @@ fn convert_to_pcm(audio_path: &Path) -> Result<Vec<f32>, String> {
     }
 
     Ok(samples)
-}
-
-/// Simple linear interpolation resampling
-fn resample(samples: &[f32], from_rate: u32, to_rate: u32) -> Vec<f32> {
-    if from_rate == to_rate {
-        return samples.to_vec();
-    }
-
-    let ratio = from_rate as f64 / to_rate as f64;
-    let new_len = (samples.len() as f64 / ratio).ceil() as usize;
-    let mut resampled = Vec::with_capacity(new_len);
-
-    for i in 0..new_len {
-        let pos = i as f64 * ratio;
-        let idx = pos.floor() as usize;
-        let frac = pos - pos.floor();
-
-        let sample = if idx + 1 < samples.len() {
-            samples[idx] * (1.0 - frac as f32) + samples[idx + 1] * frac as f32
-        } else if idx < samples.len() {
-            samples[idx]
-        } else {
-            0.0
-        };
-
-        resampled.push(sample);
-    }
-
-    resampled
 }
 
 /// Transcribe raw PCM samples directly (16kHz mono f32)
